Temporizador funcionando (aunque no para nunca :P)

+++ include/temporizadores.h
@@ -1,12 +1,14 @@
 /*-------------------------------------
 temporizadores.h
 -------------------------------------*/
+#include <nds.h>
+
 extern int tiempo;
 
 extern void HabilitarIntTemp();
 extern void DeshabilitarIntTemp();
 
-extern void prepararTemporizador(short id, int frecuencia);
-extern void iniciarTemporizador(short id);
-extern void pararTemporizador(short id);
+extern void prepararTemporizador(uint8 id, int frecuencia);
+extern void iniciarTemporizador(uint8 id);
+extern void pararTemporizador(uint8 id);
 extern void intTemporizador();
